# code/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/courses": {
      "get": {
        "operationId": "CoursesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facultyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all courses",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a course (superuser only)",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single course by ID",
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a course (superuser only)",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a course (superuser only)",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/{id}/enroll": {
      "post": {
        "operationId": "CoursesController_enroll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollStudentsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Assign/update student enrollments for a course (superuser only)",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/bulk": {
      "post": {
        "operationId": "CoursesController_bulkImport",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkImportCoursesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Bulk import courses from JSON array (superuser only)",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/api/departments": {
      "get": {
        "operationId": "DepartmentsController_findAll",
        "parameters": [
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Departments"
        ]
      },
      "post": {
        "operationId": "DepartmentsController_create",
        "parameters": [
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Departments"
        ]
      }
    },
    "/api/departments/bulk": {
      "post": {
        "operationId": "DepartmentsController_bulkCreate",
        "parameters": [
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Departments"
        ]
      }
    },
    "/api/departments/{id}": {
      "delete": {
        "operationId": "DepartmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Departments"
        ]
      }
    }
  },
  "info": {
    "title": "ENDUR API",
    "description": "\n      REST API for the **ENDUR Performance Review & Feedback Management System**.\n\n      ## Role-Based Access Control\n      All protected endpoints require the `X-Role` header with one of:\n      `superuser` | `admin` | `dean` | `hod` | `faculty` | `student`\n\n      Additionally, pass `X-User-Id` and `X-User-Name` headers to attribute audit logs correctly.\n\n      ## Authentication\n      Use `POST /api/auth/login` to verify credentials. The response contains the user object\n      which you should store client-side (session) and use to populate role headers.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "CS101"
          },
          "name": {
            "type": "string",
            "example": "Data Structures"
          },
          "facultyId": {
            "type": "string",
            "example": "F101",
            "description": "Faculty user ID"
          },
          "faculty": {
            "type": "string",
            "example": "Dr. Alan Turing"
          },
          "department": {
            "type": "string",
            "example": "Computer Science"
          },
          "type": {
            "type": "string",
            "example": "standard"
          },
          "enrolled": {
            "type": "number",
            "example": 0
          },
          "thumbnail": {
            "type": "string",
            "example": "img_read.jpg"
          }
        },
        "required": [
          "id",
          "name",
          "facultyId"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Data Structures Advanced"
          },
          "facultyId": {
            "type": "string"
          },
          "faculty": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          }
        }
      },
      "EnrollStudentsDto": {
        "type": "object",
        "properties": {
          "studentIds": {
            "example": [
              "S001",
              "S002"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "autoDept": {
            "type": "boolean",
            "example": true,
            "description": "Auto-enroll entire department"
          }
        },
        "required": [
          "studentIds"
        ]
      },
      "BulkImportCoursesDto": {
        "type": "object",
        "properties": {
          "courses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCourseDto"
            }
          }
        },
        "required": [
          "courses"
        ]
      },
...
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}